#136 - BoardList, Board 컴포넌트 분리
Components 파일에 BoardItem 과 AnswerList, AnswerItem 파일을 분리하였습니다
