+++ client/src/Pages/Boards/Board/BoardList.jsx
@@ -7,6 +7,7 @@ import MobileAside from "../Aside/MobileAside.jsx";
 import SearchBar from "../../../Components/Board/BoardList/SearchBar.jsx";
 import Dropdown from "../../../Components/Board/BoardList/Dropdown.jsx";
 import PageNation from "../../../Components/Board/BoardList/PageNation.jsx";
+import BoardListItem from "../../../Components/Board/BoardListItem/BoardListItem";
 
 //로컬 모듈
 import {
@@ -25,7 +26,7 @@ import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import { useQuery } from "@tanstack/react-query";
-import boardListStore from "../../../store/boardListStore";
+import useBoardListStore from "../../../store/useBoardListStore";
 
 export const PageWrapper = styled.div`
   display: flex;
@@ -132,67 +133,6 @@ export const BoardItem = styled.div`
   }
 `;
 
-const BoardItemContent = styled.div`
-  margin-left: 30px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-
-  @media screen and (max-width: ${breakpoint.mobile}) {
-    justify-content: flex-start;
-  }
-
-  .titleButton {
-    width: max-content;
-    margin-top: 10px;
-    font-size: ${dtFontSize.medium};
-    font-weight: 700;
-    text-align: left;
-    color: ${sub.sub900};
-    background-color: white;
-    border: none;
-    cursor: pointer;
-
-    @media screen and (max-width: ${breakpoint.mobile}) {
-      margin-top: 10px;
-      font-size: ${mbFontSize.large};
-    }
-  }
-  .contentDiv {
-    font-size: ${dtFontSize.small};
-    font-weight: 500;
-    text-align: left;
-    color: ${sub.sub700};
-
-    @media screen and (max-width: ${breakpoint.mobile}) {
-      margin-top: 10px;
-      font-size: ${mbFontSize.medium};
-    }
-  }
-`;
-
-const BoardItemCreateInfo = styled.div`
-  display: flex;
-  flex-direction: row;
-
-  @media screen and (max-width: ${breakpoint.mobile}) {
-    display: none;
-  }
-  .authorDiv {
-    font-size: ${dtFontSize.xsmall};
-    font-weight: 300;
-    color: ${sub.sub300};
-  }
-
-  .createDateDiv {
-    font-size: ${dtFontSize.xsmall};
-    font-weight: 300;
-    margin-left: 20px;
-    margin-bottom: 10px;
-    color: ${sub.sub300};
-  }
-`;
-
 const WriteButtonDiv = styled.div`
   margin-top: 30px;
   display: flex;
@@ -266,34 +206,7 @@ export default function BoardList() {
         <div className="lineDiv"></div>
         {/* boardList 가져오기 */}
         {BoardDummy.map((it) => (
-          <BoardItem key={it.id}>
-            <div className="likeDiv">
-              <div>
-                <div className="heartImageDiv">
-                  <img width={30} src={heart} alt="heart"></img>
-                </div>
-              </div>
-              <div className="heartCount">157</div>
-            </div>
-            <div className="imageDiv">
-              <img width={50} src={heart} alt="heart"></img>
-            </div>
-            <BoardItemContent>
-              <button
-                className="titleButton"
-                onClick={() => {
-                  navigate("/board/1");
-                }}
-              >
-                {it.title}
-              </button>
-              <div className="contentDiv">{it.content}</div>
-              <BoardItemCreateInfo>
-                <div className="authorDiv">{it.author}</div>
-                <div className="createDateDiv">{it.createdData}</div>
-              </BoardItemCreateInfo>
-            </BoardItemContent>
-          </BoardItem>
+          <BoardListItem key={it.id} {...it} />
         ))}
         <WriteButtonDiv>
           <WriteButton
