+++ client/src/store/useAnswerListStore.js
@@ -1,0 +1,13 @@
+import { create } from "zustand";
+
+const useAnswerListStore = create((set) => ({
+  answerList: [],
+
+  getAnswerListData(data) {
+    set((state) => ({
+      answerList: [...data],
+    }));
+  },
+}));
+
+export default useAnswerListStore;
